# api/v1/api.yaml
...
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "Dagu"
  contact:
    name: "Yota Hamada"
    url: "https://github.com/yottahmd"
  description: API for controlling and monitoring Dagu server.
  license:
    name: "GPL-3.0"
    url: "https://github.com/dagu-org/dagu/blob/main/LICENSE.md"
servers:
  - url: "{schema}://{host}/api/v1"
    description: "Dagu API server"
    variables:
      schema:
        default: http
        enum: [http, https]
      host:
        default: localhost
        description: "Host name of the server"
...
paths:
...
  /settings/agent:
    get:
      summary: "Get agent configuration"
      description: "Returns the current AI agent configuration. Requires admin role."
      operationId: "getAgentConfig"
      tags:
        - "settings"
      parameters:
        - $ref: "#/components/parameters/RemoteNode"
      responses:
        "200":
          description: "Agent configuration"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentConfigResponse"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Requires admin role"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: "Update agent configuration"
      description: "Updates the AI agent configuration. Requires admin role."
      operationId: "updateAgentConfig"
      tags:
        - "settings"
      parameters:
        - $ref: "#/components/parameters/RemoteNode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentConfigRequest"
      responses:
        "200":
          description: "Updated agent configuration"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentConfigResponse"
        "400":
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Requires admin role"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: "Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    apiToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    RemoteNode:
      name: remoteNode
      in: query
      description: name of the remote node
      required: false
      schema:
        type: string
        default: "local"
...
  schemas:
...
    Error:
      type: object
      description: "Generic error response object"
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: "Short error message"
        details:
          type: object
          description: "Additional error details"
      required:
        - code
        - message
...
    ErrorCode:
      type: string
      description: "Error code indicating the type of error"
      enum:
        - "forbidden"
        - "bad_request"
        - "not_found"
        - "internal_error"
        - "unauthorized"
        - "bad_gateway"
        - "remote_node_error"
        - "max_run_reached"
        - "not_running"
        - "already_exists"
        - "auth.unauthorized"
        - "auth.token_invalid"
        - "auth.forbidden"
        - "timeout"
...
    AgentConfigResponse:
      type: object
      description: "AI Agent configuration"
      properties:
        enabled:
          type: boolean
          description: "Whether the AI agent is enabled"
        llm:
          $ref: "#/components/schemas/AgentLLMConfig"
    AgentLLMConfig:
      type: object
      description: "LLM configuration for the AI agent"
      properties:
        provider:
          type: string
          description: "LLM provider (anthropic, openai, gemini, openrouter, local)"
        model:
          type: string
          description: "Model ID"
        apiKeyConfigured:
          type: boolean
          description: "Whether an API key is configured (actual key is not returned)"
        baseUrl:
          type: string
          description: "Custom API endpoint URL (optional)"
    UpdateAgentConfigRequest:
      type: object
      description: "Request to update AI agent configuration"
      properties:
        enabled:
          type: boolean
          description: "Whether the AI agent is enabled"
        llm:
          $ref: "#/components/schemas/UpdateAgentLLMConfig"
    UpdateAgentLLMConfig:
      type: object
      description: "LLM configuration updates"
      properties:
        provider:
          type: string
          description: "LLM provider (anthropic, openai, gemini, openrouter, local)"
        model:
          type: string
          description: "Model ID"
        apiKey:
          type: string
          description: "API key (only set if changing)"
        baseUrl:
          type: string
          description: "Custom API endpoint URL"
...
security:
  - apiToken: []
  - basicAuth: []
  - {}